third_party/blink/web_tests/fast/dom/Window/window-property-clearing-expected.txt

This page tests whether global declarations are cleared after a navigation. If the test passes, you'll see a series of PASS messages below.


Page 0:
PASS: 'x' in childWindow should be true and is.
PASS: childWindow.x should be 1 and is.
PASS: 'f' in childWindow should be true and is.
PASS: childWindow.f should be function and is.

Page 1:
PASS: 'x' in childWindow should be false and is.
PASS: childWindow.x should be undefined and is.
PASS: 'f' in childWindow should be false and is.
PASS: typeof childWindow.f should be undefined and is.
PASS: 'name' in childWindow should be true and is.
PASS: childWindow.name should be test and is.
